Extract build_readable_text into shared and reuse it in sncast script.rs
