+++ crates/forge-runner/Cargo.toml
@@ -45,3 +45,4 @@ cheatnet = { path = "../cheatnet" }
 runtime = { path = "../runtime" }
 conversions = { path = "../conversions" }
 scarb-api = { path = "../scarb-api" }
+shared = { path = "../shared" }
