+++ crates/shared/Cargo.toml
@@ -2,3 +2,7 @@
 name = "shared"
 version = "0.1.0"
 edition.workspace = true
+
+[dependencies]
+cairo-felt.workspace = true
+cairo-lang-runner.workspace = true
