+++ crates/sncast/Cargo.toml
@@ -28,6 +28,7 @@ indoc.workspace = true
 tempfile.workspace = true
 runtime = { path = "../runtime" }
 conversions = { path = "../conversions" }
+shared = { path = "../shared" }
 cairo-lang-runner.workspace = true
 cairo-lang-casm.workspace = true
 cairo-lang-sierra-to-casm.workspace = true
